+++ demo_app/lib/screens/hello_world.dart
@@ -1,4 +1,5 @@
 import 'package:demo_app/widgets/popup_menu.dart';
+import 'package:demo_app/widgets/selection_area.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 
@@ -145,7 +146,7 @@ class HelloWorldScreen extends StatelessWidget {
   const HelloWorldScreen({Key key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) => Scaffold(
+  Widget build(BuildContext context) => SelectionAreaScaffold(
         appBar: AppBar(
           title: const Text('HelloWorldScreen'),
           actions: const [
@@ -158,13 +159,7 @@ class HelloWorldScreen extends StatelessWidget {
         body: SingleChildScrollView(
           child: Padding(
             padding: const EdgeInsets.all(8.0),
-            child: HtmlWidget(
-              kHtml,
-              isSelectable: context.isSelectable,
-              key: context.key,
-              // ignore: deprecated_member_use
-              webView: true,
-            ),
+            child: HtmlWidget(kHtml, key: context.key),
           ),
         ),
       );