+++ packages/enhanced/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.9.0+1
+
+- Fix context usage across async gaps (#835)
+
+## 0.9.0
+
+- Requires Flutter 3.3 (#821)
+- Add support for `SelectionArea` (#821)
+- Breaking changes: remove support for web view `HtmlWidget` parameters (#830)
+
 ## 0.8.5
 
 - Improve `IFRAME` error & timer handling (#709)