+++ packages/fwfh_webview/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.6.2+5
+
+- Fix context usage across async gaps (#835)
+
+## 0.6.2+4
+
+- Add support for flutter_widget_from_html_core@0.9.0 (#828)
+
+## 0.6.2+3
+
+- Suppress Flutter 3 warning about `WidgetsBinding` usage
+
 ## 0.6.2+2
 
 - Improve error & timer handling (#709)